chore: define isPublicURL to determine whether the image is to be upload or not
